- add comment

+++ src/main/java/com/phuongjolly/blog/models/Comment.java
@@ -7,17 +7,18 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer userId;
+    private Long userId;
     @Lob
     private String content;
+
     @ManyToOne
     private Post post;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
